general: remove arrayvec dependency

+++ 2016/day-05/input.rs
@@ -18,7 +18,7 @@ input_params! {
 	pub struct InputParams {
 		pub num_zeros: u8 = ("NUM_ZEROS=", 5, 1_u8 .. ),
 		pub max_threads: u32 = ("MAX_THREADS=", u32::MAX, 1_u32 .. ),
-		pub password_len: usize = ("PASSWORD_LEN=", 8, 1_usize ..= 8),
+		pub password_len: u8 = ("PASSWORD_LEN=", 8, 1 ..= 8),
 		pub batch_size: usize = ("BATCH_SIZE=", 10_000, 1_usize .. ),
 	}
 }
